Merged in Cucky's branch, solved conflict

+++ source/level.cpp
@@ -25,13 +25,13 @@ void loadStageTable()
 	if (tblStream == nullptr)
 		doError();
 
-	size_t stages = SDL_RWsize(tblStream) / 200;
+	size_t stages = (size_t)SDL_RWsize(tblStream) / 200;
 
 	stageTable = static_cast<STAGE_TABLE *>(malloc(stages * 200));
 
 	if (tblStream)
 	{
-		for (int i = 0; i < stages; i++)
+		for (size_t i = 0; i < stages; i++)
 		{
 			tblStream->read(tblStream, stageTable[i].tileset, 0x20, 1);
 			tblStream->read(tblStream, stageTable[i].filename, 0x20, 1);
